+++ drivers/iommu/io-pgtable-dart.c
@@ -242,7 +242,7 @@ static int dart_map_pages(struct io_pgtable_ops *ops, unsigned long iova,
 	if (WARN_ON(pgsize != cfg->pgsize_bitmap))
 		return -EINVAL;
 
-	if (WARN_ON(paddr >> cfg->oas))
+	if (WARN_ON((u64)paddr >> cfg->oas))
 		return -ERANGE;
 
 	if (!(iommu_prot & (IOMMU_READ | IOMMU_WRITE)))
